requirements.txt
fastapi
uvicorn
python-dotenv
motor
redis
pydantic
pytest
httpx

# Optional extras (install via pip install -r requirements.txt or select subsets)
# openai
# opentelemetry-sdk
# opentelemetry-exporter-otlp
# firebase-admin# BUDDY 2.0 Production Requirements for Render Deployment
# Compatible with Python 3.13+

# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Database Dependencies
motor>=3.3.0
pymongo>=4.6.0
dnspython>=2.4.0

# Vector Database (remove pinecone for now - incompatible with Python 3.13)
# pinecone-client>=3.0.0  # Disabled - not compatible with Python 3.13

# MongoDB Integration
pymongo[srv]>=4.6.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Authentication and Security
pyjwt>=2.8.0
bcrypt>=4.1.0
cryptography>=41.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0

# AI and NLP (Core - Python 3.13 compatible)
openai>=1.3.0
sentence-transformers>=2.2.2
transformers>=4.35.0

# Audio Processing (Basic - cloud deployment compatible)
# soundfile>=0.12.0  # Disabled - may require system audio libs

# Utilities
asyncio-throttle>=1.0.0
requests>=2.31.0

# Logging and Monitoring
structlog>=23.0.0
prometheus-client>=0.19.0

# Observability - optional tracing
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0

# Date/Time
python-dateutil>=2.8.0

# JSON Processing
orjson>=3.9.0

# Background Tasks (optional for basic deployment)
# celery>=5.3.0  # Disabled for basic deployment
# redis>=5.0.0   # Disabled for basic deployment

# Cross-Platform Support
platformdirs>=4.0.0

# Production WSGI Server
gunicorn>=21.2.0

# Health Checks
psutil>=5.9.0

# Basic ML for core operations
scikit-learn>=1.3.0

# Firebase SDK
firebase-admin>=6.2.0

# For cross-device sync
websockets>=12.0.0

# Memory optimization
pympler>=0.9.0

# Async database operations (simplified)
# databases[postgresql]>=0.8.0  # Disabled for basic deployment

# Basic NLP (Python 3.13 compatible)
# spacy>=3.7.0  # Disabled - complex dependency chain

# Basic text processing
nltk>=3.8.0

# Alternative TTS (cloud-compatible)
gTTS>=2.3.0

# Development and Testing (minimal for production)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0

# NOTE: For production deployment, complex ML/audio libraries are disabled
# to ensure compatibility with Python 3.13 and cloud environments.
# Advanced features like Pinecone, spaCy, TTS can be enabled separately
# when their Python 3.13 support improves.
